requirements.txt
discord.py>=2.0.0
python-dotenv
apscheduler>=3.9.1
holidays>=0.25